used a sup for http_client and used lhttpc
I added a sup for the http_client procs for monitoring
I experimented with lhttpc; not sure if it is any better.

I may want to ditch the http_client proc and stream directly to
the http request process.

It is a bit hard to see where the slow down is but compared to velox
the performance isn't very impressive.